engine: extract esm/{utilities/histogram,layers/imageset_layer}.js
