+++ engine/esm/transpiled.js
@@ -137,10 +137,12 @@ import { MinorPlanets } from "./minor_planets.js";
 import { TileCache } from "./tile_cache.js";
 import { Tour } from "./tour.js";
 import { VideoOutputType } from "./video_output_type.js";
+import { Histogram } from "./utilities/histogram.js";
 
 import { Layer } from "./layers/layer.js";
 import { GreatCirlceRouteLayer } from "./layers/great_circle_route_layer.js";
 import { GridLayer } from "./layers/grid_layer.js";
+import { ImageSetLayer } from "./layers/imageset_layer.js";
 
 
 // wwtlib.PointType
@@ -15163,270 +15165,6 @@ var Dialog$ = {
 
 registerType("Dialog", [Dialog, Dialog$, null]);
 
-// wwtlib.Histogram
-
-export function Histogram() {
-  this.image = null;
-  this.layer = null;
-  this.tile = null;
-  this._dropDown = null;
-  this._downPosition = 0;
-  this._lowPosition = 0;
-  this._highPosition = 255;
-  this._center = 127;
-  this._ignoreNextClick = false;
-  this._dragType = 4;
-  this._updated = false;
-  this.selectedCurveStyle = 0;
-}
-
-Histogram.updateImage = function (isl, z) {
-  if (!useGlVersion2) {
-    var image = ss.safeCast(isl.get_imageSet().get_wcsImage(), FitsImageJs);
-    var Tile = tileCacheGetTile(0, 0, 0, isl.get_imageSet(), null);
-    Tile.texture2d = image.getBitmap().getTexture();
-  }
-};
-
-Histogram.updateScale = function (isl, scale, low, hi) {
-  isl.get_imageSet().get_fitsProperties().scaleType = scale;
-  isl.get_imageSet().get_fitsProperties().lowerCut = low;
-  isl.get_imageSet().get_fitsProperties().upperCut = hi;
-  if (!useGlVersion2) {
-    var image = ss.safeCast(isl.get_imageSet().get_wcsImage(), FitsImageJs);
-    var Tile = tileCacheGetTile(0, 0, 0, isl.get_imageSet(), null);
-    Tile.texture2d = image.getBitmap().getTexture();
-  }
-};
-
-Histogram.updateColorMapper = function (isl, colorMapperName) {
-  isl.get_imageSet().get_fitsProperties().colorMapName = colorMapperName;
-  if (!useGlVersion2) {
-    var image = ss.safeCast(isl.get_imageSet().get_wcsImage(), FitsImageJs);
-    var Tile = tileCacheGetTile(0, 0, 0, isl.get_imageSet(), null);
-    Tile.texture2d = image.getBitmap().getTexture();
-  }
-};
-
-var Histogram$ = {
-  close: function (e) {
-    var menu = document.getElementById('histogram');
-    var closeBtn = document.getElementById('histogramClose');
-    menu.style.display = 'none';
-    window.removeEventListener('click', ss.bind('close', this), true);
-    var image = document.getElementById('graph');
-    image.removeEventListener('mousedown', ss.bind('onPointerDown', this), false);
-    image.removeEventListener('mousemove', ss.bind('onPointerMove', this), false);
-    image.removeEventListener('mouseup', ss.bind('onPointerUp', this), false);
-    this._dropDown.removeEventListener('change', ss.bind('curveStyleSelected', this), false);
-    this._dropDown.removeEventListener('click', ss.bind('ignoreMe', this), true);
-  },
-
-  show: function (position) {
-    this.tile = tileCacheGetTile(0, 0, 0, this.layer.get_imageSet(), null);
-    var picker = document.getElementById('histogram');
-    var closeBtn = document.getElementById('histogramClose');
-    picker.style.display = 'block';
-    picker.style.left = position.x.toString() + 'px';
-    picker.style.top = position.y.toString() + 'px';
-    this.selectedCurveStyle = this.layer.get_imageSet().get_fitsProperties().scaleType;
-    this._dropDown = document.getElementById('ScaleTypePicker');
-    this._dropDown.addEventListener('change', ss.bind('curveStyleSelected', this), false);
-    this._dropDown.addEventListener('click', ss.bind('ignoreMe', this), true);
-    var canvas = document.getElementById('graph');
-    canvas.addEventListener('pointerdown', ss.bind('onPointerDown', this), false);
-    canvas.addEventListener('pointermove', ss.bind('onPointerMove', this), false);
-    canvas.addEventListener('pointerup', ss.bind('onPointerUp', this), false);
-    closeBtn.addEventListener('click', ss.bind('close', this), true);
-    this.draw();
-  },
-
-  ignoreMe: function (e) {
-    this._ignoreNextClick = true;
-  },
-
-  curveStyleSelected: function (e) {
-    this.selectedCurveStyle = this._dropDown.selectedIndex;
-    this.setUpdateTimer();
-    this.layer.get_imageSet().get_fitsProperties().scaleType = this.selectedCurveStyle;
-    this.draw();
-    this._ignoreNextClick = true;
-  },
-
-  onPointerDown: function (e) {
-    var canvas = document.getElementById('graph');
-    var x = Mouse.offsetX(canvas, e);
-    var y = Mouse.offsetY(canvas, e);
-    canvas.setPointerCapture(e.pointerId);
-    if ((Math.abs(x - this._center) < 10) && Math.abs(y - 75) < 10) {
-      this._dragType = 3;
-    } else if (Math.abs(x - this._lowPosition) < 10) {
-      this._dragType = 0;
-    } else if (Math.abs(x - this._highPosition) < 10) {
-      this._dragType = 1;
-    } else {
-      this._dragType = 2;
-      this._downPosition = Math.min(255, Math.max(0, x));
-      this.draw();
-    }
-    e.cancelBubble = true;
-  },
-
-  onPointerMove: function (e) {
-    var canvas = document.getElementById('graph');
-    var x = Mouse.offsetX(canvas, e);
-    var y = Mouse.offsetY(canvas, e);
-    switch (this._dragType) {
-      case 0:
-        this._lowPosition = Math.min(255, Math.max(0, x));
-        break;
-      case 1:
-        this._highPosition = Math.min(255, Math.max(0, x));
-        break;
-      case 2:
-        this._lowPosition = this._downPosition;
-        this._highPosition = Math.min(255, Math.max(0, x));
-        break;
-      case 3:
-        var hWidth = Math.abs(this._highPosition - this._lowPosition) / 2;
-        var adCenter = Math.min(255 - hWidth, Math.max(hWidth, x));
-        var moved = this._center - adCenter;
-        this._lowPosition -= moved;
-        this._highPosition -= moved;
-        break;
-      case 4:
-        return;
-      default:
-        break;
-    }
-    this._center = (this._lowPosition + this._highPosition) / 2;
-    this.draw();
-    var factor = (this.layer.get_imageSet().get_fitsProperties().maxVal - this.layer.get_imageSet().get_fitsProperties().minVal) / 256;
-    var low = this.layer.get_imageSet().get_fitsProperties().minVal + (this._lowPosition * factor);
-    var hi = this.layer.get_imageSet().get_fitsProperties().minVal + (this._highPosition * factor);
-    this.setUpdateTimer();
-    this.layer.get_imageSet().get_fitsProperties().upperCut = hi;
-    this.layer.get_imageSet().get_fitsProperties().lowerCut = low;
-    this.layer.get_imageSet().get_fitsProperties().scaleType = this.selectedCurveStyle;
-    e.cancelBubble = true;
-  },
-
-  onPointerUp: function (e) {
-    e.srcElement.releasePointerCapture(e.pointerId);
-    if (this._dragType !== 4) {
-      this._dragType = 4;
-      this.setUpdateTimer();
-      this._ignoreNextClick = true;
-    }
-    e.cancelBubble = true;
-  },
-
-  setUpdateTimer: function () {
-    var $this = this;
-
-    if (!useGlVersion2) {
-      setTimeout(function () {
-        $this.update();
-      }, 500);
-      this._updated = false;
-    }
-  },
-
-  update: function () {
-    if (this._updated) {
-      return;
-    }
-    if (ss.canCast(this.image, FitsImageJs)) {
-      var factor = (this.layer.get_imageSet().get_fitsProperties().maxVal - this.layer.get_imageSet().get_fitsProperties().minVal) / 256;
-      var low = this.layer.get_imageSet().get_fitsProperties().minVal + (this._lowPosition * factor);
-      var hi = this.layer.get_imageSet().get_fitsProperties().minVal + (this._highPosition * factor);
-      this.tile.texture2d = (this.image).getScaledBitmap(low, hi, this.selectedCurveStyle, 0, null).getTexture();
-    }
-    this._updated = true;
-  },
-
-  draw: function () {
-    var canvas = document.getElementById('graph');
-    var ctx = canvas.getContext('2d');
-    if (this.image != null) {
-      this.image.drawHistogram(ctx);
-    }
-    var red = 'rgba(255,0,0,255)';
-    var green = 'rgba(0,255,0,255)';
-    var blue = 'rgba(0,0,255,255)';
-    ctx.strokeStyle = red;
-    ctx.beginPath();
-    ctx.moveTo(this._lowPosition, 0);
-    ctx.lineTo(this._lowPosition, 150);
-    ctx.stroke();
-    ctx.strokeStyle = green;
-    ctx.beginPath();
-    ctx.moveTo(this._highPosition, 0);
-    ctx.lineTo(this._highPosition, 150);
-    ctx.stroke();
-    ctx.strokeStyle = blue;
-    ctx.beginPath();
-    ctx.arc(this._center, 75, 10, 0, Math.PI * 2, false);
-    ctx.closePath();
-    ctx.stroke();
-    var Curve = [];
-    switch (this.selectedCurveStyle) {
-      case 0:
-        Curve.length = 0;
-        Curve.push(Vector2d.create(this._lowPosition, 150));
-        Curve.push(Vector2d.create(this._highPosition, 0));
-        break;
-      case 1:
-        Curve.length = 0;
-        var factor = 150 / Math.log(255);
-        var diff = (this._highPosition - this._lowPosition);
-        var jump = (diff < 0) ? -1 : 1;
-        var step = Math.abs(256 / ((!diff) ? 1E-06 : diff));
-        var val = 1E-06;
-        for (var i = this._lowPosition; i !== this._highPosition; i += jump) {
-          Curve.push(Vector2d.create(i, (150 - (Math.log(val) * factor))));
-          val += step;
-        }
-        break;
-      case 2:
-        Curve.length = 0;
-        var factor = 150 / Math.pow(255, 2);
-        var diff = (this._highPosition - this._lowPosition);
-        var jump = (diff < 0) ? -1 : 1;
-        var step = Math.abs(256 / ((!diff) ? 1E-06 : diff));
-        var val = 1E-06;
-        for (var i = this._lowPosition; i !== this._highPosition; i += jump) {
-          Curve.push(Vector2d.create(i, (150 - (Math.pow(val, 2) * factor))));
-          val += step;
-        }
-        break;
-      case 3:
-        Curve.length = 0;
-        var factor = 150 / Math.sqrt(255);
-        var diff = (this._highPosition - this._lowPosition);
-        var jump = (diff < 0) ? -1 : 1;
-        var step = Math.abs(256 / ((!diff) ? 1E-06 : diff));
-        var val = 1E-06;
-        for (var i = this._lowPosition; i !== this._highPosition; i += jump) {
-          Curve.push(Vector2d.create(i, (150 - (Math.sqrt(val) * factor))));
-          val += step;
-        }
-        break;
-    }
-    if (Curve.length > 1) {
-      ctx.beginPath();
-      ctx.strokeStyle = blue;
-      ctx.moveTo(Curve[0].x, Curve[0].y);
-      for (var i = 1; i < Curve.length; i++) {
-        ctx.lineTo(Curve[i].x, Curve[i].y);
-      }
-      ctx.stroke();
-    }
-  }
-};
-
-registerType("Histogram", [Histogram, Histogram$, null]);
-
 // wwtlib.FolderDownloadAction
 
 export function FolderDownloadAction(action, loadChildFolders) {
@@ -17493,219 +17231,6 @@ export function WWTElementEvent(x, y) {
 var WWTElementEvent$ = {};
 
 registerType("WWTElementEvent", [WWTElementEvent, WWTElementEvent$, null]);
-
-// wwtlib.ImageSetLayer
-
-export function ImageSetLayer() {
-  this._imageSet$1 = null;
-  this._extension$1 = '.txt';
-  this._overrideDefaultLayer$1 = false;
-  this._loaded$1 = false;
-  Layer.call(this);
-}
-
-ImageSetLayer.create = function (set) {
-  var isl = new ImageSetLayer();
-  isl._imageSet$1 = set;
-  return isl;
-};
-
-var ImageSetLayer$ = {
-  get_imageSet: function () {
-    return this._imageSet$1;
-  },
-
-  set_imageSet: function (value) {
-    this._imageSet$1 = value;
-    return value;
-  },
-
-  get_overrideDefaultLayer: function () {
-    return this._overrideDefaultLayer$1;
-  },
-
-  set_overrideDefaultLayer: function (value) {
-    this._overrideDefaultLayer$1 = value;
-    return value;
-  },
-
-  getFitsImage: function () {
-    return ss.safeCast(this._imageSet$1.get_wcsImage(), FitsImage);
-  },
-  _isFitsImageset$1: function () {
-    var hasFitsExt = this._imageSet$1.get_extension() === '.fits';
-    return ss.canCast(this._imageSet$1.get_wcsImage(), FitsImage) || (this._imageSet$1.get_wcsImage() == null && hasFitsExt);
-  },
-
-  initializeFromXml: function (node) {
-    var imageSetNode = Util.selectSingleNode(node, 'ImageSet');
-    this._imageSet$1 = Imageset.fromXMLNode(imageSetNode);
-    if (node.attributes.getNamedItem('Extension') != null) {
-      this._extension$1 = node.attributes.getNamedItem('Extension').nodeValue;
-    }
-    if (node.attributes.getNamedItem('ScaleType') != null) {
-      this.get_imageSet().get_fitsProperties().scaleType = Enums.parse('ScaleTypes', node.attributes.getNamedItem('ScaleType').nodeValue);
-    }
-    if (node.attributes.getNamedItem('MinValue') != null) {
-      this.get_imageSet().get_fitsProperties().minVal = parseFloat(node.attributes.getNamedItem('MinValue').nodeValue);
-      this.get_imageSet().get_fitsProperties().lowerCut = (node.attributes.getNamedItem('LowerCut') != null) ? parseFloat(node.attributes.getNamedItem('LowerCut').nodeValue) : this.get_imageSet().get_fitsProperties().minVal;
-    }
-    if (node.attributes.getNamedItem('MaxValue') != null) {
-      this.get_imageSet().get_fitsProperties().maxVal = parseFloat(node.attributes.getNamedItem('MaxValue').nodeValue);
-      this.get_imageSet().get_fitsProperties().upperCut = (node.attributes.getNamedItem('UpperCut') != null) ? parseFloat(node.attributes.getNamedItem('UpperCut').nodeValue) : this.get_imageSet().get_fitsProperties().maxVal;
-    }
-    if (node.attributes.getNamedItem('ColorMapperName') != null) {
-      this.get_imageSet().get_fitsProperties().colorMapName = node.attributes.getNamedItem('ColorMapperName').nodeValue;
-    }
-    if (node.attributes.getNamedItem('OverrideDefault') != null) {
-      this._overrideDefaultLayer$1 = ss.boolean(node.attributes.getNamedItem('OverrideDefault').nodeValue);
-    }
-  },
-
-  draw: function (renderContext, opacity, flat) {
-    if (!this._loaded$1) {
-      return false;
-    }
-    renderContext.set_worldBase(renderContext.get_world());
-    renderContext.set_viewBase(renderContext.get_view());
-    renderContext.makeFrustum();
-    renderContext.drawImageSet(this._imageSet$1, this.get_opacity() * opacity * 100);
-    return true;
-  },
-
-  writeLayerProperties: function (xmlWriter) {
-    if (this._imageSet$1.get_wcsImage() != null) {
-      if (this._isFitsImageset$1()) {
-        this._extension$1 = '.fit';
-      }
-      else {
-        this._extension$1 = '.png';
-      }
-      xmlWriter._writeAttributeString('Extension', this._extension$1);
-    }
-    if (this._isFitsImageset$1()) {
-      xmlWriter._writeAttributeString('ScaleType', Enums.toXml('ScaleTypes', this._imageSet$1.get_fitsProperties().scaleType));
-      xmlWriter._writeAttributeString('MinValue', this._imageSet$1.get_fitsProperties().minVal.toString());
-      xmlWriter._writeAttributeString('MaxValue', this._imageSet$1.get_fitsProperties().maxVal.toString());
-      xmlWriter._writeAttributeString('LowerCut', this._imageSet$1.get_fitsProperties().lowerCut.toString());
-      xmlWriter._writeAttributeString('UpperCut', this._imageSet$1.get_fitsProperties().upperCut.toString());
-      if (this._imageSet$1.get_fitsProperties().colorMapName != null) {
-        xmlWriter._writeAttributeString('ColorMapperName', this._imageSet$1.get_fitsProperties().colorMapName);
-      }
-    }
-    xmlWriter._writeAttributeString('OverrideDefault', this._overrideDefaultLayer$1.toString());
-    Imageset.saveToXml(xmlWriter, this._imageSet$1, '');
-    Layer.prototype.writeLayerProperties.call(this, xmlWriter);
-  },
-
-  getTypeName: function () {
-    return 'TerraViewer.ImageSetLayer';
-  },
-
-  cleanUp: function () {
-    Layer.prototype.cleanUp.call(this);
-  },
-
-  addFilesToCabinet: function (fc) {
-    if (ss.canCast(this._imageSet$1.get_wcsImage(), FitsImage)) {
-      var fName = (this._imageSet$1.get_wcsImage()).get_filename();
-      var fileName = fc.tempDirectory + ss.format('{0}\\{1}{2}', fc.get_packageID(), this.id.toString(), this._extension$1);
-      fc.addFile(fileName, (this._imageSet$1.get_wcsImage()).sourceBlob);
-    }
-  },
-
-  getParamNames: function () {
-    return Layer.prototype.getParamNames.call(this);
-  },
-
-  getParams: function () {
-    return Layer.prototype.getParams.call(this);
-  },
-
-  setParams: function (paramList) {
-    Layer.prototype.setParams.call(this, paramList);
-  },
-
-  setImageScale: function (scaleType, min, max) {
-    console.warn('SetImageScale is considered deprecated. Use setImageScaleRaw or setImageScalePhysical instead.');
-    this.setImageScaleRaw(scaleType, min, max);
-  },
-
-  setImageScaleRaw: function (scaleType, min, max) {
-    this.get_imageSet().get_fitsProperties().lowerCut = min;
-    this.get_imageSet().get_fitsProperties().upperCut = max;
-    this.get_imageSet().get_fitsProperties().scaleType = scaleType;
-    if (ss.canCast(this._imageSet$1.get_wcsImage(), FitsImageJs)) {
-      Histogram.updateScale(this, scaleType, min, max);
-    }
-  },
-
-  setImageScalePhysical: function (scaleType, min, max) {
-    var newMin = min;
-    var newMax = max;
-    if (this._isFitsImageset$1()) {
-      newMin = (newMin - this._imageSet$1.get_fitsProperties().bZero) / this._imageSet$1.get_fitsProperties().bScale;
-      newMax = (newMax - this._imageSet$1.get_fitsProperties().bZero) / this._imageSet$1.get_fitsProperties().bScale;
-    }
-    this.setImageScaleRaw(scaleType, newMin, newMax);
-  },
-
-  setImageZ: function (z) {
-    if (this._isFitsImageset$1()) {
-      Histogram.updateImage(this, z);
-    }
-  },
-
-  get_colorMapperName: function () {
-    return this.get_imageSet().get_fitsProperties().colorMapName;
-  },
-
-  set_colorMapperName: function (value) {
-    if (ColorMapContainer.fromNamedColormap(value) == null) {
-      throw new Error('Invalid colormap name');
-    }
-    this.version++;
-    if (this._isFitsImageset$1()) {
-      if (useGlVersion2) {
-        this._imageSet$1.get_fitsProperties().colorMapName = value;
-      }
-      else {
-        Histogram.updateColorMapper(this, value);
-      }
-    }
-    return value;
-  },
-
-  get_colorMapper: function () {
-    if (this.get_imageSet().get_fitsProperties().colorMapName == null) {
-      return null;
-    } else {
-      return ColorMapContainer.fromNamedColormap(this.get_imageSet().get_fitsProperties().colorMapName);
-    }
-  },
-
-  loadData: function (tourDoc, filename) {
-    if (ss.startsWith(this._extension$1.toLowerCase(), '.fit')) {
-      var blob = tourDoc.getFileBlob(ss.replaceString(filename, '.txt', this._extension$1));
-      var fi;
-      if (useGlVersion2) {
-        fi = new FitsImage(this._imageSet$1, 'image.fit', blob, ss.bind('doneLoading', this));
-      }
-      else {
-        fi = new FitsImageJs(this._imageSet$1, 'image.fit', blob, ss.bind('doneLoading', this));
-      }
-      this._imageSet$1.set_wcsImage(fi);
-    } else {
-      this._loaded$1 = true;
-    }
-  },
-
-  doneLoading: function (wcsImage) {
-    this._loaded$1 = true;
-  }
-};
-
-registerType("ImageSetLayer", [ImageSetLayer, ImageSetLayer$, Layer]);
 
 
 // wwtlib.Object3dLayer
